+++ src/features/placeDetail/apis/placeDetailApi.ts
@@ -1,0 +1,13 @@
+import api from '../../../api/api';
+import type { DetailPlaceProps } from '../../../types/type';
+
+export const getPlaceDetails = async (
+  placeId: string,
+): Promise<DetailPlaceProps> => {
+  try {
+    const response = await api.get(`/places/${placeId}`);
+    return response.data.data;
+  } catch (err) {
+    throw err;
+  }
+};
