+++ src/features/placeDetail/place-review-card/ReviewCard.tsx
@@ -131,7 +131,7 @@ const ReviewCard = ({ review }: ReviewCardProps) => {
       <LoginModal isOpen={isLoginOpen} onClose={() => setIsLoginOpen(false)}>
         <h2 className="mb-1 text-2xl font-bold text-[#8BE34A]">로그인</h2>
         <p className="mb-4 text-xs text-gray-500">
-          이 기능을 사용하려면 로그인해주세요.
+          좋아요 기능을 사용하려면 로그인해주세요.
         </p>
         <LoginForm />
       </LoginModal>
